fix: add achievements to env.py, add not null and regenerate migration
