+++ src/backend/user-service/alembic/env.py
@@ -43,7 +43,7 @@ from shared.database import Base             # noqa: E402
 from models.user import User  # noqa: E402, F401
 
 # Tables owned by this service — Alembic will ONLY manage these.
-_SERVICE_TABLES = {"users"}
+_SERVICE_TABLES = {"users", "achievements", "user_achievements", "user_xp"}
 
 # Each service gets its own version-tracking table so they don't collide.
 _VERSION_TABLE = "alembic_version_user"
