--- src/backend/user-service/alembic/versions/20260408_bc9aaeac902a_add_achievements_table.py
@@ -1,66 +1,0 @@
-"""add_achievements_table
-
-Revision ID: bc9aaeac902a
-Revises: c1d2e3f4a5b6
-Create Date: 2026-04-08 04:52:02.884833
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bc9aaeac902a'
-down_revision: Union[str, None] = 'c1d2e3f4a5b6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'achievements',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('key', sa.String(length=50), nullable=False),
-        sa.Column('name', sa.String(length=100), nullable=False),
-        sa.Column('description', sa.Text(), nullable=False),
-        sa.Column('icon', sa.String(length=100), nullable=True),
-        sa.PrimaryKeyConstraint('id'),
-        sa.UniqueConstraint('key')
-    )
-
-    # Create user_achievements table (junction table)
-    op.create_table(
-        'user_achievements',
-        sa.Column('user_id', sa.Integer(), nullable=False),
-        sa.Column('achievement_id', sa.Integer(), nullable=False),
-        sa.Column('earned_at',
-                  sa.TIMESTAMP(timezone=True),
-                  server_default=sa.text('now()')),
-        sa.PrimaryKeyConstraint('user_id', 'achievement_id'),
-        sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-        sa.ForeignKeyConstraint(['achievement_id'], ['achievements.id'], )
-    )
-
-    # Create user_xp table
-    op.create_table(
-        'user_xp',
-        sa.Column('user_id', sa.Integer(), nullable=False),
-        sa.Column('xp', sa.Integer(), server_default='0'),
-        sa.Column('level', sa.Integer(),
-                  sa.Computed('xp / 100 + 1')),
-        sa.PrimaryKeyConstraint('user_id'),
-        sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE')
-    )
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('user_xp')
-    op.drop_table('user_achievements')
-    op.drop_table('achievements')
-    # ### end Alembic commands ###
